// types/creative.ts
export interface Creative {
  creative_id: string
  creative_name: string
  tags: string
  country: string
  ad_network: string
  os: string
  campaign: string
  ad_group: string
  ipm: number
  ctr: number
  spend: number
  impressions: number
  clicks: number
  cpm: number
  cost_per_click: number
  cost_per_install: number
  installs: number
}